+++ lib/tcx/mixins.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
 require_relative 'mixins/distance_meters'
+require_relative 'mixins/average_speed'